Test.Shelley.Spec.Ledger.Examples

Documentation

data CHAINExample

Constructors

CHAINExample

Fields

startState ∷ ChainState

State to start testing with

newBlock ∷ Block

Block to run chain state transition system on

intendedResult ∷ Either [[PredicateFailure CHAIN]] ChainState

type of fatal error, if failure expected and final chain state if success expected

ex1 ∷ CHAINExample

Wraps example all together.

ex2A ∷ CHAINExample

ex2B ∷ CHAINExample

Wrap plain example

ex2C ∷ CHAINExample

Example 2C with standard decay.

ex2Cbis ∷ CHAINExample

Example 2C with no decay.

ex2Cter ∷ CHAINExample

Example 2C with full refund.

ex2Cquater ∷ CHAINExample

Example 2C with instant decay.

ex2D ∷ CHAINExample

ex2E ∷ CHAINExample

ex2F ∷ CHAINExample

ex2G ∷ CHAINExample

ex2H ∷ CHAINExample

ex2I ∷ CHAINExample

ex2J ∷ CHAINExample

ex2K ∷ CHAINExample

ex2L ∷ CHAINExample

ex3A ∷ CHAINExample

ex3B ∷ CHAINExample

ex3C ∷ CHAINExample

ex4A ∷ CHAINExample

ex4B ∷ CHAINExample

ex5A ∷ CHAINExample

ex5B ∷ CHAINExample

ex5C ∷ CHAINExample

Example 5C - Instantaneous rewards that overrun the available reserves

ex5D' ∷ Either [[PredicateFailure CHAIN]] ChainState

ex6A ∷ CHAINExample

ex6BExpectedNES ∷ NewEpochState

ex6BPoolParams ∷ Map (KeyHash StakePool) PoolParams

test5D ∷ Assertion

Tests that after getting instantaneous rewards, creating the update and then applying the update, Alice's key is actually registered, the key deposit value deducted and the remaining value credited as reward.

alicePay ∷ KeyPair Payment

aliceStake ∷ KeyPair Staking

aliceAddr ∷ Addr

bobPay ∷ KeyPair Payment

bobStake ∷ KeyPair Staking

bobAddr ∷ Addr

carlPay ∷ KeyPair Payment

carlStake ∷ KeyPair Staking

carlAddr ∷ Addr

dariaPay ∷ KeyPair Payment

dariaStake ∷ KeyPair Staking

dariaAddr ∷ Addr

coreNodeSKG ∷ Int → SignKeyDSIGN

blockEx1 ∷ Block

Null initial block. Just records the Byron hash, and contains no transactions.

blockEx2A ∷ Block

blockEx2B ∷ Block

blockEx2C ∷ Block

Example 2C - process an empty block in the next epoch so that the (empty) reward update is applied and a stake snapshot is made.

blockEx2D ∷ Block

blockEx2E ∷ Block

Example 2E - create the first non-empty pool distribution by creating a block in the third epoch of this running example.

blockEx2F ∷ Block

blockEx2G ∷ Block

Example 2G - create an empty block in the next epoch to prepare the way for the first non-trivial reward update

blockEx2H ∷ Block

Example 2H - create the first non-trivial reward update

blockEx2I ∷ Block

Example 2I - apply the first non-trivial reward update

blockEx2J ∷ Block

blockEx2K ∷ Block

blockEx2L ∷ Block

Example 2L - reap a stake pool

blockEx3A ∷ Block

blockEx3B ∷ Block

blockEx3C ∷ Block

Example 3C - Adopt protocol parameter update

blockEx4A ∷ Block

blockEx4B ∷ Block

Example 4B - New genesis key delegation updated from future delegations

blockEx5A ∷ Block

blockEx5B ∷ Block

blockEx5F ∷ Block

blockEx5F' ∷ Block

blockEx5F'' ∷ Block

txEx2A ∷ Tx

txEx2B ∷ Tx

txEx2D ∷ Tx

txEx2J ∷ Tx

txEx2K ∷ Tx

txEx3A ∷ Tx

txEx3B ∷ Tx

txEx4A ∷ Tx

txEx5A ∷ Tx

txEx5B ∷ Tx

Example 5B - Instantaneous rewards with insufficient core node signatures

txEx5F ∷ Tx

txEx5F' ∷ Tx

txEx5F'' ∷ Tx

unsafeMkUnitInterval ∷ Rational → UnitInterval

You vouch that argument is in [0; 1].