Test.Shelley.Spec.Ledger.Generator.Constants

Documentation

minNumGenInputs ∷ Int

minimal number of transaction inputs to select

maxNumGenInputs ∷ Int

maximal number of transaction inputs to select

frequencyRegKeyCert ∷ Int

Relative frequency of generated credential registration certificates

frequencyRegPoolCert ∷ Int

Relative frequency of generated pool registration certificates

frequencyDelegationCert ∷ Int

Relative frequency of generated delegation certificates

frequencyGenesisDelegationCert ∷ Int

Relative frequency of generated genesis delegation certificates

frequencyDeRegKeyCert ∷ Int

Relative frequency of generated credential de-registration certificates

frequencyRetirePoolCert ∷ Int

Relative frequency of generated pool retirement certificates

frequencyMIRCert ∷ Int

Relative frequency of generated MIR certificates

frequencyScriptCredReg ∷ Int

Relative frequency of script credentials in credential registration certificates

frequencyKeyCredReg ∷ Int

Relative frequency of key credentials in credential registration certificates

frequencyScriptCredDeReg ∷ Int

Relative frequency of script credentials in credential de-registration certificates

frequencyKeyCredDeReg ∷ Int

Relative frequency of key credentials in credential de-registration certificates

frequencyScriptCredDelegation ∷ Int

Relative frequency of script credentials in credential delegation certificates

frequencyKeyCredDelegation ∷ Int

Relative frequency of key credentials in credential delegation certificates

frequencyTxUpdates ∷ Int

Relative frequency of Prototol/Application Updates in a transaction

minGenesisUTxOouts ∷ Int

minimal number of genesis UTxO outputs

maxGenesisUTxOouts ∷ Int

maximal number of genesis UTxO outputs

maxCertsPerTx ∷ Word64

maximal number of certificates per transaction

maxNumKeyPairs ∷ Word64

maximal numbers of generated keypairs

minGenesisOutputVal ∷ Integer

minimal coin value for generated genesis outputs

maxGenesisOutputVal ∷ Integer

maximal coin value for generated genesis outputs

numBaseScripts ∷ Int

Number of base scripts from which multi sig scripts are built.

frequencyNoWithdrawals ∷ Int

Relative frequency that a transaction does not include any reward withdrawals

frequencyAFewWithdrawals ∷ Int

Relative frequency that a transaction includes a small number of reward withdrawals, bounded by maxAFewWithdrawals.

maxAFewWithdrawals ∷ Int

Maximum number of reward withdrawals that counts as a small number.

frequencyPotentiallyManyWithdrawals ∷ Int

Relative frequency that a transaction includes any positive number of reward withdrawals

minSlotTrace ∷ Int

Minimal slot for CHAIN trace generation.

maxSlotTrace ∷ Int

Maximal slot for CHAIN trace generation.

frequencyLowMaxEpoch ∷ Word64

Lower bound of the MaxEpoch protocol parameter

maxMinFeeA ∷ Natural

maxMinFeeB ∷ Natural