Test.Shelley.Spec.Ledger.Generator.Core

Documentation

data AllPoolKeys

Constructors

AllPoolKeys

Fields

cold ∷ KeyPair

vrf ∷ (SignKeyVRF, VerKeyVRF)

hot ∷ [(KESPeriod, (SKeyES, VKeyES))]

hk ∷ KeyHash

newtype NatNonce

We provide our own nonces to mkBlock, which we then wish to recover as the output of the VRF functions. In general, however, we just derive them from a natural. Since the nonce is a hash, we do not want to recover it to find a preimage. In testing, therefore, we just wrap the raw natural, which we then encode into the fake VRF implementation.

Constructors

NatNonce Natural

coreNodeVKG ∷ Int → VKeyGenesis

findPayKeyPairAddr ∷ Addr → Map AnyKeyHash KeyPair → KeyPair

Find first matching key pair for address. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayKeyPairCred ∷ Credential → Map AnyKeyHash KeyPair → KeyPair

Find first matching key pair for a credential. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayScriptFromCred ∷ Credential → MultiSigPairs → (MultiSig, MultiSig)

Find first matching script for a credential.

findStakeScriptFromCred ∷ Credential → MultiSigPairs → (MultiSig, MultiSig)

Find first matching script for a credential.

findPayScriptFromAddr ∷ Addr → MultiSigPairs → (MultiSig, MultiSig)

Find first matching script for address.

genBool ∷ Gen Bool

genCoin ∷ Integer → Integer → Gen Coin

genCoinList ∷ Integer → Integer → Int → Int → Gen [Coin]

Generates a list of Coin values of length between lower and upper and with values between minCoin and maxCoin.

genInteger ∷ Integer → Integer → Gen Integer

genNatural ∷ Natural → Natural → Gen Natural

Generator for a natural number between lower and upper

genWord64 ∷ Word64 → Word64 → Gen Word64

Generator for a Word64 between lower and upper

genTxOut ∷ [Addr] → Gen [TxOut]

Generates a list of coins for the given Addr and produced a TxOut for each Addr

Note: we need to keep the initial utxo coin sizes large enough so that when we simulate sequences of transactions, we have enough funds available to include certificates that require deposits.

genUtxo0 ∷ Int → Int → Gen UTxO

genesisAccountState ∷ AccountState

Account with empty treasury

genesisDelegs0 ∷ Map GenKeyHash KeyHash

increasingProbabilityAt ∷ Gen a → (a, a) → Gen a

Generate values the given distribution in 90% of the cases, and values at the bounds of the range in 10% of the cases.

This can be used to generate enough extreme values. The exponential and linear distributions provided by hedgehog will generate a small percentage of these (0-1%).

numCoreNodes ∷ Word64

coreKeyPairs ∷ [CoreKeyPair]

coreNodeKeys ∷ [(CoreKeyPair, AllPoolKeys)]

traceKeyPairs ∷ KeyPairs

Constant list of KeyPairs intended to be used in the generators.

traceKeyPairsByStakeHash ∷ Map KeyHash KeyPair

traceKeyHashMap ∷ Map AnyKeyHash KeyPair

Mapping from key hash to key pair

traceVRFKeyPairs ∷ [(SignKeyVRF, VerKeyVRF)]

A pre-populated space of VRF keys for use in the generators.

traceVRFKeyPairsByHash ∷ Map VRFKeyHash (SignKeyVRF, VerKeyVRF)

traceMSigScripts ∷ MultiSigPairs

Multi-Sig Scripts based on the traceKeyPairs key pairs

traceMSigCombinations ∷ MultiSigPairs → MultiSigPairs

Combine a list of multisig pairs into hierarchically structured multi-sig scripts, list must have at least length 3. Be careful not to call with too many pairs in order not to create too many of the possible combinations.

someKeyPairs ∷ Int → Int → Gen KeyPairs

Select between _lower_ and _upper_ keys from traceKeyPairs

someScripts ∷ Int → Int → Gen MultiSigPairs

Select between _lower_ and _upper_ scripts from the possible combinations of the first numBaseScripts multi-sig scripts of traceMSigScripts.

pickStakeKey ∷ KeyPairs → Gen VKey

Select one random verification staking key from list of pairs of KeyPair.

toAddr ∷ Crypto crypto ⇒ (KeyPair Regular crypto, KeyPair Regular crypto) → Addr crypto

toCred ∷ Crypto crypto ⇒ KeyPair Regular crypto → Credential crypto

zero ∷ UnitInterval

unitIntervalToNatural ∷ UnitInterval → Natural

Try to map the unit interval to a natural number. We don't care whether this is surjective. But it should be right inverse to fromNatural - that is, one should be able to recover the UnitInterval value used here.

mkBlock

Arguments

∷ HashHeader
Hash of previous block

→ AllPoolKeys
All keys in the stake pool

→ [Tx]
Transactions to record

→ SlotNo
Current slot

→ BlockNo
Block numberchain lengthchain "difficulty"

→ Nonce
EpochNo nonce

→ NatNonce
Block nonce

→ UnitInterval
Praos leader value

→ Natural
Period of KES (key evolving signature scheme)

→ Natural
KES period of key registration

→ OCert
Operational certificate

→ Block

mkOCert ∷ AllPoolKeys → Natural → KESPeriod → OCert

getKESPeriodRenewalNo ∷ AllPoolKeys → KESPeriod → Integer

tooLateInEpoch ∷ SlotNo → Bool

True if the given slot is within the last `2 * slotsPrior` slots of the current epoch.