Test.Shelley.Spec.Ledger.Generator.Delegation

Documentation

genDCert ∷ (HasCallStack, Era era) ⇒ Constants → KeySpace era → PParams era → AccountState → DPState era → SlotNo → Gen (Maybe (DCert era, CertCred era))

Occasionally generate a valid certificate

Returning Nothing indicates a failure to generate a value, usually due to lack of available values from the pre-populated (e.g. key) spaces. A Just represents a successfully generated value.

Different generators return witnesses that are either genesis or regular keys.

Note: we register keys and pools more often than deregistering/retiring them, and we generate more delegations than registrations of keys/pools.

data CertCred h

Constructors

CoreKeyCred [GenesisKeyPair h]

StakeCred (KeyPair Staking h)

PoolCred (KeyPair StakePool h)

ScriptCred (MultiSig h, MultiSig h)

DelegateCred [KeyPair GenesisDelegate h]

NoCred