Test.Shelley.Spec.Ledger.Generator.Presets

Description

Pre-generated items to use in traces.

Functions in this module make specific assumptions about the sets of keys involved, and thus cannot be used as generic generators.

Documentation

coreNodeKeys ∷ Constants → [(CoreKeyPair, AllPoolKeys)]

keySpace ∷ Constants → KeySpace

Example keyspace for use in generators

genEnv ∷ GenEnv

Example generator environment, consisting of default constants and an corresponding keyspace.

genUtxo0 ∷ Constants → Gen UTxO

genesisDelegs0 ∷ Constants → Map GenKeyHash KeyHash