Test.Shelley.Spec.Ledger.Generator.Trace.Chain

Documentation

lastByronHeaderHash ∷ ∀ proxy era. Era era ⇒ proxy era → HashHeader (Crypto era)

The first block of the Shelley era will point back to the last block of the Byron era. For our purposes we can bootstrap the chain by just coercing the value. When this transition actually occurs, the consensus layer will do the work of making sure that the hash gets translated across the fork

mkGenesisChainState ∷ ∀ era a. ShelleyTest era ⇒ Gen (Value era) → Constants → IRC (CHAIN era) → Gen (Either a (ChainState era))

mkOCertIssueNos ∷ GenDelegs h → Map (KeyHash 'BlockIssuer h) Natural

registerGenesisStaking ∷ ∀ c. ShelleyTest c ⇒ ShelleyGenesisStaking c → ChainState c → ChainState c

Orphan instances

(ShelleyTest era, GetLedgerView era, ApplyBlock era, STS (CHAIN era), BaseM (CHAIN era) ~ ShelleyBase, STS (LEDGERS era), BaseM (LEDGERS era) ~ ShelleyBase, Environment (CHAIN era) ~ (), State (CHAIN era) ~ ChainState era, Signal (CHAIN era) ~ Block era, Environment (LEDGERS era) ~ LedgersEnv era, State (LEDGERS era) ~ LedgerState era, Signal (LEDGERS era) ~ Seq (Tx era), STS (LEDGER era), BaseM (LEDGER era) ~ ShelleyBase, Environment (LEDGER era) ~ LedgerEnv era, State (LEDGER era) ~ (UTxOState era, DPState era), Signal (LEDGER era) ~ Tx era, Mock (Crypto era)) ⇒ HasTrace (CHAIN era) (GenEnv era)

Instance details

Associated Types

type BaseEnv (CHAIN era)

Methods

interpretSTS ∷ BaseEnv (CHAIN era) → BaseM (CHAIN era) a → a

envGen ∷ GenEnv era → Gen (Environment (CHAIN era))

sigGen ∷ GenEnv era → Environment (CHAIN era) → State (CHAIN era) → Gen (Signal (CHAIN era))

shrinkSignal ∷ Signal (CHAIN era) → [Signal (CHAIN era)]