Test.Shelley.Spec.Ledger.Generator.Trace.Ledger

Documentation

genTx_ ∷ LedgerEnv → (UTxOState, DPState) → Gen Tx

mkGenesisLedgerState ∷ IRC LEDGER → Gen (Either a (UTxOState, DPState))

Generate initial state for the LEDGER STS using the STS environment.

Note: this function must be usable in place of applySTS and needs to align with the signature 'RuleContext sts -> Gen (Either [[PredicateFailure sts]] (State sts))'. To achieve this we (1) use 'IRC LEDGER' (the "initial rule context") instead of simply LedgerEnv and (2) always return Right (since this function does not raise predicate failures).

Orphan instances

HasTrace LEDGERS Word64

Instance details

Associated Types

type BaseEnv LEDGERS ∷ Type

Methods

interpretSTS ∷ BaseEnv LEDGERS → BaseM LEDGERS a → a

envGen ∷ Word64 → Gen (Environment LEDGERS)

sigGen ∷ Word64 → Environment LEDGERS → State LEDGERS → Gen (Signal LEDGERS)

shrinkSignal ∷ Signal LEDGERS → [Signal LEDGERS]

HasTrace LEDGER Word64

Instance details

Associated Types

type BaseEnv LEDGER ∷ Type

Methods

interpretSTS ∷ BaseEnv LEDGER → BaseM LEDGER a → a

envGen ∷ Word64 → Gen (Environment LEDGER)

sigGen ∷ Word64 → Environment LEDGER → State LEDGER → Gen (Signal LEDGER)

shrinkSignal ∷ Signal LEDGER → [Signal LEDGER]