Test.Shelley.Spec.Ledger.Generator.Utxo

Documentation

genTx ∷ (HasCallStack, Mock c) ⇒ GenEnv c → LedgerEnv → (UTxOState c, DPState c) → Gen (Tx c)

Generates a transaction in the context of the LEDGER STS environment and state.

A generated transaction may not have sufficient spending balance and need to be discarded. In this case we retry the generator, up to genTxRetries times, before failing hard with an error.

Note: the spending balance emerges from inputs, refund withdrawals, certificate deposits and fees (which in turn depend on number of inputs, outputs, witnesses, metadata etc.). It's hard to avoid this completely, but in practice it is relatively easy to calibrate the generator Constants so that there is sufficient spending balance.