Test.Shelley.Spec.Ledger.NonTraceProperties.Mutator

Description

Module Description : Generators for mutating data.

This module implements a mutation based approach for generating invalid data. Input values are mutated depending on their value.

Documentation

mutateNat ∷ Natural → Natural → Natural → Gen Natural

Mutator that combines the identity, range selector and small change mutator in a random choice.

mutateCoin ∷ Natural → Natural → Coin → Gen Coin

Mutator for Coin values, based on mutation of the contained value field.

mutateTx ∷ Tx → Gen Tx

Mutator of Tx which mutates the contained transaction

mutateTxBody ∷ TxBody → Gen TxBody

Mutator for Transaction which mutates the set of inputs and the set of unspent outputs.

mutateDCert ∷ KeyPairs → DPState → DCert → Gen DCert

Mutator for delegation certificates. A RegKey and DeRegKey select randomly a key fomr the supplied list of keypairs. A RetirePool certificate mutates the epoch and the key of the certificate. A RegPool certificate mutates the staking key, the pool's cost and margin. A Delegate certificates selects randomly keys for delegator and delegatee from the supplied list of keypairs.

getAnyStakeKey ∷ KeyPairs → Gen (VKey Staking)

Select one random verification staking key from list of pairs of KeyPair.