Test.Shelley.Spec.Ledger.NonTraceProperties.Validity

Documentation

data ValidationError

Validation errors represent the failures of a transaction to be valid for a given ledger state.

Constructors

BadInputs
The transaction inputs are not valid.

Expired SlotNo SlotNo
The transaction has expired

RetirementCertExpired SlotNo SlotNo
Pool Retirement Certificate expired

FeeTooSmall Coin Coin
The transaction fee is too small

ValueNotConserved Coin Coin
Value is not conserved

IncorrectRewards
Unknown reward account

InvalidWitness
One of the transaction witnesses is invalid.

MissingWitnesses
The transaction does not have the required witnesses.

InputSetEmpty
Missing Replay Attack Protection, at least one input must be spent.

StakeKeyAlreadyRegistered
A stake key cannot be registered again.

StakeKeyNotRegistered
A stake key must be registered to be used or deregistered.

StakeDelegationImpossible
The stake key to which is delegated is not known.

StakePoolNotRegisteredOnKey
Stake pool not registered for key, cannot be retired.

UnknownValidationError
Other error (for conversion from STS errors that have no correspondance here)

data Validity

The validity of a transaction, where an invalid transaction is represented by list of errors.

Constructors

Valid

Invalid [ValidationError]

data LedgerValidation

A ledger validation state consists of a ledger state t and the list of validation errors that occurred from a valid s to reach t.

Constructors

LedgerValidation [ValidationError] (LedgerState ConcreteCrypto)

validInputs ∷ Crypto crypto ⇒ TxBody crypto → UTxOState crypto → Validity

Determine if the inputs in a transaction are valid for a given ledger state.

current ∷ Crypto crypto ⇒ TxBody crypto → SlotNo → Validity

Determine if the transaction has expired

validNoReplay ∷ Crypto crypto ⇒ TxBody crypto → Validity

Determine if the input set of a transaction consumes at least one input, else it would be possible to do a replay attack using this transaction.

validFee ∷ ∀ crypto. Crypto crypto ⇒ PParams → Tx crypto → Validity

Determine if the fee is large enough

preserveBalance ∷ Crypto crypto ⇒ StakePools crypto → StakeCreds crypto → PParams → TxBody crypto → UTxOState crypto → Validity

Determine if the balance of the ledger state would be effected in an acceptable way by a transaction.

correctWithdrawals ∷ RewardAccounts crypto → RewardAccounts crypto → Validity

Determine if the reward witdrawals correspond to the rewards in the ledger state

validRuleUTXO ∷ Crypto crypto ⇒ RewardAccounts crypto → StakePools crypto → StakeCreds crypto → PParams → SlotNo → Tx crypto → UTxOState crypto → Validity

verifiedWits' ∷ (Crypto crypto, DSignable crypto (Hash crypto (TxBody crypto))) ⇒ Tx crypto → Validity

Given a ledger state, determine if the UTxO witnesses in a given transaction are correct.

enoughWits ∷ Crypto crypto ⇒ Tx crypto → GenDelegs crypto → UTxOState crypto → Validity

Given a ledger state, determine if the UTxO witnesses in a given transaction are sufficient. We check that there are not more witnesses than inputs, if several inputs from the same address are used, it is not strictly necessary to include more than one witness.

validRuleUTXOW ∷ (Crypto crypto, DSignable crypto (Hash crypto (TxBody crypto))) ⇒ Tx crypto → GenDelegs crypto → LedgerState crypto → Validity

validTx ∷ (Crypto crypto, DSignable crypto (Hash crypto (TxBody crypto))) ⇒ Tx crypto → GenDelegs crypto → SlotNo → PParams → LedgerState crypto → Validity