Test.Shelley.Spec.Ledger.PreSTSGenerator

Documentation

utxoSize ∷ UTxO → Int

Returns the number of entries of the UTxO set.

utxoMap ∷ UTxO → Map TxIn TxOut

Extract the map in an UTxO.

genNonEmptyAndAdvanceTx ∷ Gen (KeyPairs, Natural, Coin, LedgerState, [Tx], Either [ValidationError] LedgerState)

Generator of a non-emtpy ledger genesis state and a random number of transactions applied to it. Returns the amount of accumulated fees, the initial ledger state and the final ledger state or the validation error if an invalid transaction has been generated.

genNonEmptyAndAdvanceTx' ∷ Gen (KeyPairs, Natural, Coin, LedgerState, [Tx], LedgerValidation)

Mutated variant of above, collects validation errors in LedgerValidation.

genNonemptyGenesisState ∷ Gen LedgerState

Generator of a non-empty genesis ledger state, i.e., at least one valid address and non-zero UTxO.

genStateTx ∷ Gen (LedgerState, Natural, Coin, Tx, LedgerValidation)

genValidStateTx ∷ Gen (LedgerState, Natural, Coin, Tx, LedgerState)

genValidStateTxKeys ∷ Gen (LedgerState, Natural, Coin, Tx, LedgerState, KeyPairs)

genDelegationData ∷ KeyPairs → EpochNo → Gen DCert

genDelegation ∷ KeyPairs → DPState → Gen Delegation

genDCertDelegate ∷ KeyPairs → DPState → Gen DCert

genKeyPairs ∷ Int → Int → Gen KeyPairs

Generates a list of '(pay, stake)' key pairs.

asStateTransition ∷ SlotNo → PParams → LedgerState → Tx → Coin → Either [ValidationError] LedgerState

In the case where a transaction is valid for a given ledger state, apply the transaction as a state transition function on the ledger state. Otherwise, return a list of validation errors.

asStateTransition' ∷ SlotNo → PParams → LedgerValidation → Tx → Coin → LedgerValidation

Apply transition independent of validity, collect validation errors on the way.