Test.Shelley.Spec.Ledger.Utils

Documentation

assertAll ∷ (MonadTest m, Show a, Eq a) ⇒ (a → Bool) → [a] → m ()

mkSeedFromWords ∷ (Word64, Word64, Word64, Word64, Word64) → Seed

Construct a seed from a bunch of Word64s

We multiply these words by some extra stuff to make sure they contain enough bits for our seed.

mkCertifiedVRF ∷ ToCBOR a ⇒ WithResult a → SignKeyVRF → CertifiedVRF a

For testing purposes, create a VRF value

epochFromSlotNo ∷ SlotNo → EpochNo

evolveKESUntil

Arguments

∷ SignKeyKES

→ KESPeriod
Current KES period

→ KESPeriod
Target KES period

→ Maybe SignKeyKES

Try to evolve KES key until specific KES period is reached, given the current KES period.

slotFromEpoch ∷ EpochNo → SlotNo

mkKeyPair ∷ (Word64, Word64, Word64, Word64, Word64) → (SignKeyDSIGN, VKey kr)

For testing purposes, generate a deterministic key pair given a seed.

mkGenKey ∷ (Word64, Word64, Word64, Word64, Word64) → (SignKeyDSIGN, VKeyGenesis)

For testing purposes, generate a deterministic genesis key pair given a seed.

mkKESKeyPair ∷ (Word64, Word64, Word64, Word64, Word64) → (SignKeyKES, VerKeyKES)

For testing purposes, generate a deterministic KES key pair given a seed.

mkVRFKeyPair ∷ (Word64, Word64, Word64, Word64, Word64) → (SignKeyVRF, VerKeyVRF)

For testing purposes, generate a deterministic VRF key pair given a seed.

mkAddr ∷ (KeyPair Payment, KeyPair Staking) → Addr

runShelleyBase ∷ ShelleyBase a → a

testGlobals ∷ Globals

maxKESIterations ∷ Word64

unsafeMkUnitInterval ∷ Rational → UnitInterval

You vouch that argument is in [0; 1].

slotsPerKESIteration ∷ Word64

maxLLSupply ∷ Coin