Test.Shelley.Spec.Ledger.Generator.Core

Documentation

data AllPoolKeys

Constructors

AllPoolKeys

Fields

cold ∷ KeyPair

vrf ∷ (SignKeyVRF, VerKeyVRF)

hot ∷ [(KESPeriod, (SKeyES, VKeyES))]

hk ∷ KeyHash

data GenEnv

Generator environment.

Constructors

GenEnv

Fields

geKeySpace ∷ KeySpace

geConstants ∷ Constants

data KeySpace

Collection of all keys which are required to generate a trace.

These are the _only_ keys which should be involved in the trace.

Constructors

KeySpace_

Fields

ksCoreNodes ∷ [(CoreKeyPair, AllPoolKeys)]

ksKeyPairs ∷ KeyPairs

ksKeyPairsByHash ∷ Map AnyKeyHash KeyPair

ksKeyPairsByStakeHash ∷ Map KeyHash KeyPair

ksMSigScripts ∷ MultiSigPairs

ksVRFKeyPairs ∷ [(SignKeyVRF, VerKeyVRF)]

ksVRFKeyPairsByHash ∷ Map VRFKeyHash (SignKeyVRF, VerKeyVRF)

pattern KeySpace ∷ [(CoreKeyPair, AllPoolKeys)] → KeyPairs → MultiSigPairs → [(SignKeyVRF, VerKeyVRF)] → KeySpace

newtype NatNonce

We provide our own nonces to mkBlock, which we then wish to recover as the output of the VRF functions. In general, however, we just derive them from a natural. Since the nonce is a hash, we do not want to recover it to find a preimage. In testing, therefore, we just wrap the raw natural, which we then encode into the fake VRF implementation.

Constructors

NatNonce Natural

findPayKeyPairAddr ∷ HasCallStack ⇒ Addr → Map AnyKeyHash KeyPair → KeyPair

Find first matching key pair for address. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayKeyPairCred ∷ HasCallStack ⇒ Credential → Map AnyKeyHash KeyPair → KeyPair

Find first matching key pair for a credential. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayScriptFromCred ∷ HasCallStack ⇒ Credential → MultiSigPairs → (MultiSig, MultiSig)

Find first matching script for a credential.

findStakeScriptFromCred ∷ HasCallStack ⇒ Credential → MultiSigPairs → (MultiSig, MultiSig)

Find first matching script for a credential.

findPayScriptFromAddr ∷ HasCallStack ⇒ Addr → MultiSigPairs → (MultiSig, MultiSig)

Find first matching script for address.

genBool ∷ HasCallStack ⇒ Gen Bool

genCoin ∷ HasCallStack ⇒ Integer → Integer → Gen Coin

genCoinList ∷ HasCallStack ⇒ Integer → Integer → Int → Int → Gen [Coin]

Generates a list of Coin values of length between lower and upper and with values between minCoin and maxCoin.

genInteger ∷ HasCallStack ⇒ Integer → Integer → Gen Integer

genNatural ∷ HasCallStack ⇒ Natural → Natural → Gen Natural

Generator for a natural number between lower and upper

genWord64 ∷ HasCallStack ⇒ Word64 → Word64 → Gen Word64

Generator for a Word64 between lower and upper

genTxOut ∷ HasCallStack ⇒ Constants → [Addr] → Gen [TxOut]

Generates a list of coins for the given Addr and produced a TxOut for each Addr

Note: we need to keep the initial utxo coin sizes large enough so that when we simulate sequences of transactions, we have enough funds available to include certificates that require deposits.

increasingProbabilityAt ∷ HasCallStack ⇒ Gen a → (a, a) → Gen a

Generate values the given distribution in 90% of the cases, and values at the bounds of the range in 10% of the cases.

This can be used to generate enough extreme values. The exponential and linear distributions provided by hedgehog will generate a small percentage of these (0-1%).

mkScriptsFromKeyPair ∷ HasCallStack ⇒ (KeyPair, KeyPair) → (MultiSig, MultiSig)

pickStakeKey ∷ HasCallStack ⇒ KeyPairs → Gen VKey

Select one random verification staking key from list of pairs of KeyPair.

toAddr ∷ Crypto crypto ⇒ (KeyPair Regular crypto, KeyPair Regular crypto) → Addr crypto

toCred ∷ Crypto crypto ⇒ KeyPair Regular crypto → Credential crypto

zero ∷ HasCallStack ⇒ UnitInterval

unitIntervalToNatural ∷ HasCallStack ⇒ UnitInterval → Natural

Try to map the unit interval to a natural number. We don't care whether this is surjective. But it should be right inverse to fromNatural - that is, one should be able to recover the UnitInterval value used here.

mkBlock

Arguments

∷ HasCallStack

⇒ HashHeader
Hash of previous block

→ AllPoolKeys
All keys in the stake pool

→ [Tx]
Transactions to record

→ SlotNo
Current slot

→ BlockNo
Block numberchain lengthchain "difficulty"

→ Nonce
EpochNo nonce

→ NatNonce
Block nonce

→ UnitInterval
Praos leader value

→ Natural
Period of KES (key evolving signature scheme)

→ Natural
KES period of key registration

→ OCert
Operational certificate

→ Block

mkOCert ∷ HasCallStack ⇒ AllPoolKeys → Natural → KESPeriod → OCert

getKESPeriodRenewalNo ∷ HasCallStack ⇒ AllPoolKeys → KESPeriod → Integer

tooLateInEpoch ∷ HasCallStack ⇒ SlotNo → Bool

True if the given slot is within the last `2 * slotsPrior` slots of the current epoch.

mkKeyPair ∷ (Word64, Word64, Word64, Word64, Word64) → (SKey, VKey)

For testing purposes, generate a deterministic key pair given a seed.

mkKeyPairs ∷ HasCallStack ⇒ Word64 → (KeyPair, KeyPair)

mkGenKey ∷ (Word64, Word64, Word64, Word64, Word64) → (SKey, VKeyGenesis)

For testing purposes, generate a deterministic genesis key pair given a seed.

mkMSigScripts ∷ HasCallStack ⇒ KeyPairs → MultiSigPairs

Multi-Sig Scripts based on the given key pairs

mkMSigCombinations ∷ HasCallStack ⇒ MultiSigPairs → MultiSigPairs

Combine a list of multisig pairs into hierarchically structured multi-sig scripts, list must have at least length 3. Be careful not to call with too many pairs in order not to create too many of the possible combinations.

genesisAccountState ∷ HasCallStack ⇒ AccountState

Account with empty treasury