Test.Shelley.Spec.Ledger.Generator.Delegation

Documentation

genDCert ∷ HasCallStack ⇒ Constants → KeyPairs → MultiSigPairs → [CoreKeyPair] → VrfKeyPairs → Map (KeyHash Staking) (KeyPair Staking) → PParams → DPState → SlotNo → Gen (Maybe (DCert, CertCred))

Occasionally generate a valid certificate

Returning Nothing indicates a failure to generate a value, usually due to lack of available values from the pre-populated (e.g. key) spaces. A Just represents a successfully generated value.

Different generators return witnesses that are either genesis or regular keys.

Note: we register keys and pools more often than deregistering/retiring them, and we generate more delegations than registrations of keys/pools.

data CertCred

Constructors

CoreKeyCred [CoreKeyPair]

KeyCred (KeyPair Staking)

ScriptCred (MultiSig, MultiSig)

NoCred