Test.Shelley.Spec.Ledger.Generator.Trace.Chain

Documentation

lastByronHeaderHash ∷ HashHeader

The first block of the Shelley era will point back to the last block of the Byron era. For our purposes we can bootstrap the chain by just coercing the value. When this transition actually occurs, the consensus layer will do the work of making sure that the hash gets translated across the fork

mkGenesisChainState ∷ Constants → IRC CHAIN → Gen (Either a ChainState)

mkOCertIssueNos ∷ GenDelegs → Map (KeyHash BlockIssuer) Natural

Orphan instances

HasTrace CHAIN GenEnv

Instance details

Associated Types

type BaseEnv CHAIN ∷ Type

Methods

interpretSTS ∷ BaseEnv CHAIN → BaseM CHAIN a → a

envGen ∷ GenEnv → Gen (Environment CHAIN)

sigGen ∷ GenEnv → Environment CHAIN → State CHAIN → Gen (Signal CHAIN)

shrinkSignal ∷ Signal CHAIN → [Signal CHAIN]