Test.Shelley.Spec.Ledger.Generator.Utxo

Documentation

genTx ∷ ∀ era. (HasCallStack, ShelleyTest era, Mock (Crypto era)) ⇒ GenEnv era → LedgerEnv era → (UTxOState era, DPState era) → Gen (Tx era)

Generates a transaction in the context of the LEDGER STS environment and state.

A generated transaction may not have sufficient spending balance and need to be discarded. In that case we try to compute a Delta, that when added (applyDelta) to the transaction, repairs it. The repair is made by adding additional inputs from which more Ada can flow into the fee. If that doesn't fix it, we add add more inputs to the Delta. Experience shows that this converges quite quickly (in traces we never saw more than 3 iterations).

data Delta era

Collect additional inputs (and witnesses and keys and scripts) to make the transaction balance.

Constructors

Delta

Fields

dfees ∷ Coin

extraInputs ∷ Set (TxIn era)

extraWitnesses ∷ WitnessSet era

change ∷ TxOut era

deltaVKeys ∷ [KeyPair 'Witness (Crypto era)]

deltaScripts ∷ [(MultiSig era, MultiSig era)]

showBalance ∷ (ShelleyTest era, HasField "inputs" (TxBody era) (Set (TxIn era)), HasField "outputs" (TxBody era) (StrictSeq (TxOut era)), HasField "wdrls" (TxBody era) (Wdrl era), HasField "txfee" (TxBody era) Coin, HasField "certs" (TxBody era) (StrictSeq (DCert era))) ⇒ LedgerEnv era → UTxOState era → DPState era → Tx era → String