Test.Shelley.Spec.Ledger.Generator.Utxo

Documentation

data GenTxException

Constructors

GenTxExceptionInsufficientSpendingBalance !InsufficientSpendingBalanceInfo
Every transaction that the generator attempted had fees that exceeded the available balance.

genTx ∷ (HasCallStack, Era era, Mock (Crypto era)) ⇒ GenEnv era → LedgerEnv era → (UTxOState era, DPState era) → Gen (Tx era)

Generates a transaction in the context of the LEDGER STS environment and state.

A generated transaction may not have sufficient spending balance and need to be discarded. In this case we retry the generator, up to genTxRetries times, before failing hard with an error.

Note: the spending balance emerges from inputs, refund withdrawals, certificate deposits and fees (which in turn depend on number of inputs, outputs, witnesses, metadata etc.). It's hard to avoid this completely, but in practice it is relatively easy to calibrate the generator Constants so that there is sufficient spending balance.

tryGenTx ∷ (HasCallStack, Era era, Mock (Crypto era)) ⇒ GenEnv era → LedgerEnv era → (UTxOState era, DPState era) → Gen (Either GenTxException (Tx era))

splitCoin ∷ Coin → Integer → (Coin, Coin)