Test.Shelley.Spec.Ledger.Generator.Utxo

Documentation

genTx ∷ LedgerEnv → (UTxOState, DPState) → KeyPairs → Map AnyKeyHash KeyPair → MultiSigPairs → [(CoreKeyPair, AllPoolKeys)] → Map KeyHash KeyPair → Gen Tx

Generate a new transaction in the context of the LEDGER STS environment and state.

Selects unspent outputs and spends the funds on a some valid addresses. Also generates valid certificates.